+++ binance/um_futures/account.py
@@ -780,3 +780,45 @@ def commission_rate(self, symbol: str, **kwargs):
     params = {"symbol": symbol, **kwargs}
 
     return self.sign_request("GET", url_path, params)
+
+
+def download_transactions_asyn(self, startTime: int, endTime: int, **kwargs):
+    """
+    |
+    | **Get Download Id For Futures Transaction History (USER_DATA)**
+
+    :API endpoint: ``GET /fapi/v1/income/asyn``
+    :API doc: https://binance-docs.github.io/apidocs/futures/en/#get-download-id-for-futures-transaction-history-user_data
+
+    :parameter startTime:  int
+    :parameter endTime:  int
+    :parameter recvWindow: optional int
+    |
+    """
+
+    check_required_parameter(startTime, "startTime")
+    check_required_parameter(endTime, "endTime")
+    url_path = "/fapi/v1/income/asyn"
+    params = {"startTime": startTime, "endTime": endTime, **kwargs}
+
+    return self.sign_request("GET", url_path, params)
+
+
+def aysnc_download_info(self, downloadId: str, **kwargs):
+    """
+    |
+    | **Get Futures Transaction History Download Link by Id (USER_DATA)**
+
+    :API endpoint: ``GET /fapi/v1/income/asyn/id``
+    :API doc: https://binance-docs.github.io/apidocs/futures/en/#get-futures-transaction-history-download-link-by-id-user_data
+
+    :parameter downloadId:  string
+    :parameter recvWindow: optional int
+    |
+    """
+
+    check_required_parameter(downloadId, "downloadId")
+    url_path = "/fapi/v1/income/asyn/id"
+    params = {"downloadId": downloadId, **kwargs}
+
+    return self.sign_request("GET", url_path, params)