+++ examples/websocket/um_futures/user_data.py
@@ -3,7 +3,7 @@
 import time
 import logging
 from binance.lib.utils import config_logging
-from binance.cm_futures import CMFutures
+from binance.um_futures import UMFutures
 from binance.websocket.um_futures.websocket_client import UMFuturesWebsocketClient
 
 config_logging(logging, logging.DEBUG)
@@ -14,7 +14,7 @@ def message_handler(message):
 
 
 api_key = ""
-client = CMFutures(api_key)
+client = UMFutures(api_key)
 response = client.new_listen_key()
 
 logging.info("Listen key : {}".format(response["listenKey"]))