+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 3.1.0 - 2022-06-02
+
+### Add
+
+- New endpoint `GET /fapi/v1/income/asyn` to get Download Id For Futures Transaction History
+- New endpoint `GET /fapi/v1/income/asyn/id` to get Futures Transaction History Download Link by Id
+
+
 ## 3.0.0 - 2022-05-26
 
 - Refactor the connector